requirements.txt
Flask
ldap3
Flask-Login
Flask-Session
Flask-WTF
python-dotenv